+++ src/com/aic/aicdetactor/fragment/PartItemMeasureTemperatureFragment.java
@@ -175,10 +175,10 @@ private EditText mEditTextValue;
 //    }
 	  //临时生成随机的三维坐标数据及温度数据。
     void genRandom_temperation(){    
-    	float max_temperation=300;
-    	float MAX = 200;
-    	float MID = 100;
-    	float LOW = 0;
+    	double max_temperation=300;
+    	double MAX = 200;
+    	double MID = 100;
+    	double LOW = 0;
 		
     	MAX = super.mPartItemData.Up_Limit;
     	MID = super.mPartItemData.Middle_Limit;
